Change all require for require_relative and create new files for the classes
