+++ lib/give_quote.rb
@@ -1,0 +1,7 @@
+class GiveQuote
+  include Quotes
+  def self.random_quote
+    values = QUOTES.values
+    values[rand(values.size)]
+  end
+end
