+++ lib/hello_text.rb
@@ -1,0 +1,5 @@
+class HelloText
+  def self.hello
+    "Hey! You can see the commands if you enter the word 'help' :sunglasses:"
+  end
+end
